Refactor favicon link in index.html and fix file path
Update the favicon link in index.html to use the correct file path "/BJB-SOCIAL-LOGO.png" instead of "././public/BJB-SOCIAL-LOGO.png". This ensures that the favicon is displayed correctly on the website.

Fix styling issue in JobPostSkeleton component

In the JobPostSkeleton component, update the styling to improve the layout and visual appearance. This includes adjusting the size and positioning of various elements to create a more consistent and visually appealing skeleton loading effect.

Update font-family in index.css

Update the font-family in index.css to use "Poppins" as the primary font for the entire website. This ensures a consistent and visually pleasing typography across all pages.

Update image file paths in LoginPage and SignUpPage

Update the image file paths in LoginPage and SignUpPage to use the correct file path "/BJB-SOCIAL-LOGO.png" instead of "./../public/BJB-SOCIAL-LOGO.png". This ensures that the profile image is displayed correctly on the login and signup pages.

Improve search functionality in HomePage

Refactor the search functionality in HomePage to improve the user experience. This includes updating the state management for the search input and handling the search input change event more efficiently.

Refactor JobPost component

Refactor the JobPost component to use a unique dialog ID based on the job ID. This ensures that each job post has a unique dialog ID, which is important for proper functioning of the component.